% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_5pl_model.R
\name{alt_5pl_model}
\alias{alt_5pl_model}
\title{Calculate a 5 parameter logistic regression model with some start values from another model}
\usage{
alt_5pl_model(
  list,
  n_iter = 1000,
  weights = c("lower", "upper", "none"),
  FI = NULL,
  model_name = NULL,
  ...
)
}
\arguments{
\item{list}{one list entry as derived from plexr::read_bioplex}

\item{n_iter}{number of iterations for nls.multstart::nls_multstart}

\item{weights}{on which end of the regression does good fitting to standards
is required most (e.g. where do the measured values lie), lower range or upper range
or no weights (none); only used if FI is NULL.}

\item{FI}{provide custom weights instead of the easy-to-use weights parameter}

\item{model_name}{if NULL, just the model is returned; if model_name is not NULL
the model is added under that name to list and list is returned}

\item{...}{additional parameters passed to nls.multstart::nls_multstart; e.g.
control = nls.lm.control(maxiter = 1024, maxfev=10000); or supp_errors = "Y"}
}
\value{

}
\description{
nls.multstart::nls_multstart is used to calculate a 5 parameter logistic regression.
Multiparametric regression requires to guess starting values. This guessing is done
based on standard values (aa, dd) and on a previous regression made by the bioplex
biomanager (cc, dd, gg).
}
