% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/five_par_log_regress.R
\name{five_par_log_regress}
\alias{five_par_log_regress}
\title{5-parameter logistic regression prediction}
\usage{
five_par_log_regress(nls_model, Conc = NULL, FI = NULL)
}
\arguments{
\item{nls_model}{non-linear least squares regression model}

\item{Conc}{concentration values}

\item{FI}{fluorescence intensities}
}
\value{

}
\description{
Alternative to stats::predict. Provide concentration (Conc) or Fluorescence intensity (FI)
to have the other one returned according to the model.
}
\details{
If FI is provided, dd + (aa - dd) / ((1 + (Conc / cc)^bb))^gg is used;
if Conc is provided dd + (aa - dd) / ((1 + (Conc / cc)^bb))^gg is used.
Alternatively stats::predict(nls_model, newdata = data.frame(Conc = 1:10))
}
