% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragment_trace_plot.R
\name{fragment_trace_plot}
\alias{fragment_trace_plot}
\title{Standard plot of nucleic acid trace data from Fragment- or Bio-Analyzer}
\usage{
fragment_trace_plot(
  FA_data,
  electropherogram_entry = "electropherogram",
  quality_entry = "quality",
  x_breaks = c(15, 500, 1500, 3000, 6000),
  theme = ggplot2::theme_bw(),
  theme_args = list(axis.text.x = element_text(angle = 45, hjust = 1, vjust = 1),
    panel.grid.minor = element_blank(), strip.background = element_rect(fill =
    "cornsilk")),
  facetting = NULL,
  x_lims = c(5, NA)
)
}
\arguments{
\item{FA_data}{list of nucleic acid trace data read with plexr::read_fragmentanalyzer}

\item{electropherogram_entry}{name of the electropherogram list entry}

\item{quality_entry}{name of the quality list entry}

\item{x_breaks}{breaks on x-axis}

\item{theme}{ggplot2 theme}

\item{theme_args}{arguments to modify the theme}

\item{facetting}{how to facet the plot; see example}

\item{x_lims}{x-axis limits}
}
\value{
a ggplot2 plot object
}
\description{
Create a standardized plot for nucleid acid trace data. The electropherogram
table is joined with quality table. Hence, RQN values or concentration or R28S_18S (ratio 28S/18S) can be
used as a facetting variable. Combinations are possible: RQN_R28S_18S,
RQN_conc, R28S_18S_conc, RQN_R28S_18S_conc.
}
\examples{
\dontrun{
FA_files <- list.files(FA_folder, pattern = "\\\\.csv", full.names = T)
FA_data <- plexr::read_fragmentanalyzer(FA_files)
fragment_trace_plot(FA_data, facetting = ggplot2::facet_wrap(vars(chemical, Buffer, DTT_mM, RQN), scales = "free_y"))
}
}
