% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bioplex.R
\name{read_bioplex}
\alias{read_bioplex}
\title{Read data from Bioplex Managers standard output xlsx}
\usage{
read_bioplex(file, sheets = NULL, rows = NULL)
}
\arguments{
\item{file}{path to the xlsx file (export_per_analyte)}

\item{sheets}{names of sheets to consider; if NULL all sheets are considered}

\item{rows}{rows which contain non-aggregated measured values including the header;
e.g. 59:160 or c(59, 160); if NULL the range is guessed by occurence of "Analyte"
in the first column of worksheets}
}
\value{

}
\description{
Get standard values, sample values and parameters from standard curve as calculated by Bioplex Manager.
This function was written based a file from Bioplex Manager version 4.
Every sheet in the xlsx represents one analyte from the assay.
Reading the xlsx file is done by openxlsx::read.xlsx.
}
