% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fragmentanalyzer.R
\name{read_fragmentanalyzer}
\alias{read_fragmentanalyzer}
\title{Read csv output files from Fragment Analyzer into data frames}
\usage{
read_fragmentanalyzer(
  files,
  Electropherogram_file_pattern = "Electropherogram",
  QualityTable_file_pattern = "Quality Table",
  PeakTable_file_pattern = "Peak Table",
  filter_names = c("Ladder", "Samp.[[:digit:]]{1,2}"),
  return = c("data.frame", "list")
)
}
\arguments{
\item{files}{full path to all files}

\item{Electropherogram_file_pattern}{pattern to identify electropherogram files within files}

\item{QualityTable_file_pattern}{pattern to identify Quality Table files within files}

\item{PeakTable_file_pattern}{pattern to identify Peak Table files within files}

\item{filter_names}{vector of sample name patterns to remove, can be regular expressions; set to NULL or empty vector to not filter any sample}

\item{return}{which type to return, list of tables for each file (list); or one data.frame for all (data.frame)}
}
\value{
list of lists or list of data.frames
}
\description{
Read csv output files from Fragment Analyzer into data frames
}
\examples{
\dontrun{
# get all file paths
files <- list.files("parent_dir", recursive = T, full.names = T, pattern = "\\\\.csv")
FA_data <- plexr::read_fragmentanalyzer(files = files)
}
}
