% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_magpix.R
\name{read_magpix}
\alias{read_magpix}
\title{Read csv file from MAGPIX assay}
\usage{
read_magpix(file_path)
}
\arguments{
\item{file_path}{character, path to the csv file}
}
\value{
list of all table parts and a processed data frame of Count, Median and Result
}
\description{
Read the csv file which is the standard output of MAGPIX.
}
\examples{
\dontrun{
read_magpix('path_to_your_file.csv')
}
}
