% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nanodrop.R
\name{read_nanodrop}
\alias{read_nanodrop}
\title{Read reports and spectra files from NanoDrop into tidy tables}
\usage{
read_nanodrop(reports, spectra)
}
\arguments{
\item{reports}{vector of paths to report files (tsv)}

\item{spectra}{vector of paths to spectra files (tsv)}
}
\value{
a list of length 2; index one is report only, index two is report
with spectra values added
}
\description{
The tsv files from Nanodrop are not very handy in R. Several conversions
are necessary (column names, character to numeric etc.) which is time consuming.
Also, in order to plot spectra the respective table from Nandrop requires quite
some wrangling. This function tries to solve these tasks elegantly.
}
