% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_magpix.R
\name{report_magpix}
\alias{report_magpix}
\title{Standard report on magpix measurement}
\usage{
report_magpix(
  df,
  count_threshold = 30,
  meta_data,
  group_col = "group",
  out_folder
)
}
\arguments{
\item{df}{the data frame returned by read_magpix; filter this df to focus on certain analytes or samples only}

\item{count_threshold}{minimal bead count to include in the}

\item{meta_data}{a data frame of meta_data; should contain a column named well_id for joining and a column named group_col (next argument) for grouping observations}

\item{group_col}{column name in meta_data with values to group observations}

\item{out_folder}{path to the folder to save plots and tables to}
}
\value{
a data frame of plots and tables may also be saved to out_folder if provided
}
\description{
Standard report on magpix measurement
}
\examples{
\dontrun{
# get the path to the parent folder of a script
wd <- dirname(rstudioapi::getActiveDocumentContext()$path)
report_magpix(df = read_magpix('path_to_your_file.csv'),
count_threshold = 30,
meta_data = openxlsx::read.xlsx('path_to_meta_data_xlsx'),
out_folder = file.path(wd, "report"))
}
}
